Add drawSettingView.js
